backend/requirements.txt
Flask
python-dotenv
langchain
langchain-huggingface
langchain-community
requests
